Move encryption function to a separate file
